+++ CHANGELOG.md
@@ -115,7 +115,7 @@ That just about summarizes the major features in Alpha 16. If you'd like to know
   - Status effects no longer show above party members.
   - Larger party member status windows.
   - Gradient background behind party member status windows.
-  - Party member status windows now "slide in".
+  - Party member status windows now slide in and out.
   - Pixelated battlebacks.
 - All encounters can be run away from.
 - Infernal Wand now deals +20% fire damage.